+++ unity/Assets/Maroon/reusableScripts/ExperimentParameters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7214fc94f95e57f4bb3a4281bdbcacec
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
