+++ unity/Assets/Maroon/reusableScripts/ExperimentParameters/ExperimentParameters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09f9dc3074b79b246bfb7703caf237a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
