+++ unity/Assets/Maroon/reusableScripts/ExperimentParameters/Maroon.ReusableScripts.ExperimentParameters.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "Maroon.ReusableScripts.ExperimentParameters",
+    "rootNamespace": "",
+    "references": [
+        "GUID:56f15dd651d5d3a44bbfc9ca89629921"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
