+++ unity/Assets/Maroon/reusableScripts/ExperimentParameters/Maroon.ReusableScripts.ExperimentParameters.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 566f6fdebd119a54485118cb24dec5cb
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
