+++ unity/Assets/Maroon/scenes/experiments/3DMotionSimulation/Scripts/Maroon.Experiments.3DMotionSimulation.asmdef
@@ -2,7 +2,6 @@
     "name": "Maroon.Experiments.3DMotionSimulation",
     "rootNamespace": "",
     "references": [
-        "GUID:20a3f6c7643c02b4f91ee2b86822b6a7",
         "GUID:ccf513c02a7c0734f97bf699ec9fcc75",
         "GUID:f9b0aa2a96663864693b204ee0977e65",
         "GUID:c4aed1b7b59d62c48ab35dda9c36ff44",
@@ -15,7 +14,8 @@
         "GUID:339607d959a71f14eadd1d72a99e0287",
         "GUID:56f15dd651d5d3a44bbfc9ca89629921",
         "GUID:9d2d2b17ea939f447a2e35b833b82ffa",
-        "GUID:9f64e54542c8a1d4ea36f9126abf642c"
+        "GUID:9f64e54542c8a1d4ea36f9126abf642c",
+        "GUID:566f6fdebd119a54485118cb24dec5cb"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
