+++ unity/Assets/Maroon/scenes/experiments/3DMotionSimulation/Scripts/ThreeDimensionalMotionParameters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d61a1d3bec3de84497f54aa8815b7e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
